Refactor preferences dialog
